Added Classes for working with FotoFly specific metadata which is stored in an XMPStruct.   - IFotoFlyMetadata (Interface that declares all properties)   - FotoFlyMetadata (Class for storing data)   - WpfFotoFlyMetadata (Class for retreiving and setting data)   - FotoFlyQueries (Static class defining all the XMP Queries)
Updated BitmapMetadataHelper
 - Added IsQueryOfType to validate return type is expected type
 - Added CopyQuery
 - Added MoveQuery
 
Updated WpfMetadata
 - Updated Iso to use IsQueryOfType before running GetQuery
 - Updated GpsMeasureMode to use IsQueryOfType before running GetQuery
 
Ran JpgPhoto.ReadMetadata() on 25,000 photos
 - Fixed all 11 failures with the the fixes to WpfMetadata noted above.

Finished Merging in Geotagging Code.
 - Added GeotaggingUnitTest
 - Added BingMapsResolver to wrap Bing's Geotagging APIs
 - Added GoogleMapsResolver to wrap Google's Geotagging APIs
 - Added GeoCache for storing cached metadata queries
 - Added GpsTrackResolver to lookup GPS coordinates in a GPS Track
 - Added Reference to BingMapsForEnterprise WebService

Passed all 15 Unit Tests

+++ FotoFly/MetadataQueries/ExifQueries.cs
@@ -67,7 +67,7 @@ namespace FotoFly
 
         // Short. Indicates the ISO Speed and ISO Latitude of the camera or input device as specified
         // in ISO 12232. 
-        public static readonly MetdataQuery IsoSpeedRating = new MetdataQuery("/app1/ifd/exif/subifd:{uint=34855}", typeof(Int16));
+        public static readonly MetdataQuery IsoSpeedRating = new MetdataQuery("/app1/ifd/exif/subifd:{uint=34855}", typeof(UInt16));
 
         // Ascii. The date and time when the original image data was generated. For a digital still camera
         // the date and time the picture was taken are recorded. 
@@ -181,7 +181,7 @@ namespace FotoFly
         
         // Indicates the Gps measurement mode. '2' means two-dimensional measurement and '3' means
         // three-dimensional
-        public static readonly MetdataQuery GpsMeasureMode = new MetdataQuery("/app1/ifd/Gps/subifd:{uint=10}", typeof(Int16));
+        public static readonly MetdataQuery GpsMeasureMode = new MetdataQuery("/app1/ifd/Gps/subifd:{uint=10}", typeof(Int32));
 
         // Indicates the GPS DOP (data degree of precision). An HDOP value is written during
         // two-dimensional measurement, and PDOP during three-dimensional measurement.
